feedback edits

+++ install-aeneas.cmd
@@ -4,13 +4,12 @@
 call :check_Permissions
 if "%1" == "setupncheck" call :setupncheck
 
-echo this installer uses Chocolatey package manager for Windows
-echo first check if it is installed
+echo This installer method uses Chocolatey package manager for Windows
 call where /Q choco.exe
 if "%errorlevel%" == "0" (
   @echo Chocolatey is already installed
   ) else (
-  @echo installing Chocolatey
+  @echo   installing Chocolatey
   @powershell -NoProfile -ExecutionPolicy Bypass -Command "iex ((new-object net.webclient).DownloadString('https://chocolatey.org/install.ps1'))" && PATH "%PATH%;%ALLUSERSPROFILE%\chocolatey\bin"
   set errorlevel=0
   )
@@ -26,9 +25,9 @@ if "%errorlevel%" == "0" (
   @echo.
 where espeak /q
 if "%errorlevel%" == "0" (
-  echo espeak already installed
+  echo eSpeak already installed
   ) else (
-  echo download and install espeak
+  echo download and install eSpeak
   choco install espeak -s "%cd%" -f
 )
   @echo.
@@ -39,7 +38,8 @@ rem from https://download.microsoft.com/download/7/9/6/796EF2E4-801B-4FC4-AB28-B
 @echo Checking for Python 2.7  
 where python /q
 if "%errorlevel%" == "0" (
-  @echo Python is already installed Check next line reports "Python 2.7.xx"
+  @echo   Python is already installed. 
+  @echo   Check that next line reports "Python 2.7.xx"
   python --version
   @echo.
   ) else (
@@ -58,12 +58,15 @@ if exist "C:\Users\mcquayi\AppData\Local\Programs\Common\Microsoft\Visual C++ fo
 
 set pip=C:\tools\python2\Scripts\pip.exe
 if exist "%pip%"  (
-  echo Pip in found
-  echo starting pip Python package installs
-  %pip% install BeautifulSoup
-  %pip% install lxml
-  %pip% numpy-1.9.3+mkl-cp27-none-win_amd64.whl
-  %pip% install scikits.audiolab-0.11.0-cp27-none-win_amd64.whl
+  echo Checking if required Python packages are installed
+  if not exist "C:\tools\python2\Lib\site-packages\BeautifulSoup*dist-info" %pip% install BeautifulSoup
+  if exist "C:\tools\python2\Lib\site-packages\BeautifulSoup*dist-info" echo   BeautifulSoup installed
+  if not exist "C:\tools\python2\Lib\site-packages\lxml" %pip% install lxml
+  if exist "C:\tools\python2\Lib\site-packages\lxml" echo   lxml installed
+  if not exist "C:\tools\python2\Lib\site-packages\numpy" %pip% numpy-1.9.3+mkl-cp27-none-win_amd64.whl
+  if exist "C:\tools\python2\Lib\site-packages\numpy" echo   numpy installed
+  if not exist "C:\tools\python2\Lib\site-packages\scikits.audiolab*dist-info" %pip% install scikits.audiolab-0.11.0-cp27-none-win_amd64.whl
+  if exist "C:\tools\python2\Lib\site-packages\scikits.audiolab*dist-info" echo   scikits.audiolab installed
 
 ) else (
   echo Pip not found at %pip%
@@ -75,11 +78,13 @@ goto :eof
 
 :check_Permissions
     rem from Ben Hooper on http://stackoverflow.com/questions/4051883/batch-script-how-to-check-for-admin-rights
-    echo Administrative permissions required. Detecting permissions...
+    echo Administrative permissions required. 
+    echo   Detecting permissions...
 
     net session >nul 2>&1
     if %errorLevel% == 0 (
         echo Success: Administrative permissions confirmed.
+        @echo.
     ) else (
         echo Failure: Current permissions inadequate.
         pause >nul
